Ocultar carreras con encuestas asignadas F.

+++ app/Http/Controllers/CarreraController.php
@@ -161,40 +161,49 @@ class CarreraController extends Controller
     public function destroy($id)
     {
         $carrera = Carrera::find($id);
-        $objetivos = $carrera->objetivos;
-        foreach ($objetivos as $objetivo) {
-            foreach ($objetivo->aspectos as $aspecto){
-                foreach($aspecto->preguntas as $pregunta){
-                    $pregunta->delete();
+
+        if ($carrera->noBorrar == True) {
+            // dd('no se puede borrar');
+            $carrera->oculto = True;
+            $carrera->save();
+        }else{
+            // dd('si se puede borrar');
+            $objetivos = $carrera->objetivos;
+            foreach ($objetivos as $objetivo) {
+                foreach ($objetivo->aspectos as $aspecto){
+                    foreach($aspecto->preguntas as $pregunta){
+                        $pregunta->delete();
+                    }
                 }
+                $relacion = ObjetivoAspecto::where('objetivo_educacional_id', $objetivo->id)->get();
+                $relacion->each->delete();
+                $aspectos = $objetivo->aspectos;
+                $aspectos->each->delete();
             }
-            $relacion = ObjetivoAspecto::where('objetivo_educacional_id', $objetivo->id)->get();
-            $relacion->each->delete();
-            $aspectos = $objetivo->aspectos;
-            $aspectos->each->delete();
+    
+            $relacionObj = ObjetivoEducacional::where('idCarrera', $id)->get();
+            $relacionObj->each->delete();
+    
+            $relacionUserCarrera = UsuarioCarrera::where('carrera_id', $id)->get();
+            $relacionUserCarrera->each->delete();
+    
+            // $objetivo=ObjetivoEducacional::findorFail($id);
+            // $aspectos = $objetivo->aspectos;
+            // foreach($aspectos as $aspecto){
+            //     foreach($aspecto->preguntas as $pregunta){
+            //         $pregunta->delete();
+            //     }
+            // }
+            // $aspectos->each->preguntas->each->delete();
+            // $relacion = ObjetivoAspecto::where('objetivo_educacional_id', $id)->get();
+            // $relacion->each->delete();
+            // $aspectos->each->delete();
+    
+    
+    
+            $carrera->delete();
         }
 
-        $relacionObj = ObjetivoEducacional::where('idCarrera', $id)->get();
-        $relacionObj->each->delete();
-
-        $relacionUserCarrera = UsuarioCarrera::where('carrera_id', $id)->get();
-        $relacionUserCarrera->each->delete();
-
-        // $objetivo=ObjetivoEducacional::findorFail($id);
-        // $aspectos = $objetivo->aspectos;
-        // foreach($aspectos as $aspecto){
-        //     foreach($aspecto->preguntas as $pregunta){
-        //         $pregunta->delete();
-        //     }
-        // }
-        // $aspectos->each->preguntas->each->delete();
-        // $relacion = ObjetivoAspecto::where('objetivo_educacional_id', $id)->get();
-        // $relacion->each->delete();
-        // $aspectos->each->delete();
-
-
-
-        $carrera->delete();
 
         return redirect()->route('carreras.index');
     }
